requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
streamlit